JSDoc example for translatePlatformKeys, rename translatePlatformKey
